Updated Course Instructor Controller View Main spring2022.txt and users.txt

+++ src/SIS/CourseStudent.java
@@ -31,7 +31,7 @@ public class CourseStudent {
         return grade;
     }
 
-    public void setGrade(int grade) {
+    public void setGrade(double grade) {
         this.grade = grade;
     }
 }
